refactor(sdk-coin-cosmos): removing addition file and fix imports
TICKET: COIN-5601
