+++ modules/sdk-coin-cosmos/test/unit/transactionBuilder/transferBuilder.ts
@@ -1,6 +1,6 @@
-import { TransactionType } from '@bitgo/sdk-core';
-import { fromBase64, toHex } from '@cosmjs/encoding';
 import should from 'should';
+import { fromBase64, toHex } from '@cosmjs/encoding';
+import { TransactionType } from '@bitgo/sdk-core';
 import { getAvailableTestCoins, getBuilderFactory, getTestData } from '../../testUtils';
 
 describe('Cosmos Transfer Builder', function () {
